fix: 構造体の型変更

+++ repository/post-repo.go
@@ -3,10 +3,7 @@ package repository
 import (
 	"context"
 	"log"
-	// "strings"
-
 	"../entity"
-	// "google.golang.org/api/iterator"
 
 	"cloud.google.com/go/firestore"
 )
@@ -70,7 +67,7 @@ func (*repo) FindAll() ([]entity.Post, error) {
 		    return nil, err
 		}
 		post := entity.Post {
-			ID: doc.Data()["ID"].(int),
+			ID: doc.Data()["ID"].(int64),
 			Title: doc.Data()["Title"].(string),
 			Text: doc.Data()["Text"].(string),
 		}
